requirements.txt
absl-py==0.9.0
astor==0.8.1
backcall==0.1.0
certifi==2019.11.28
cycler==0.10.0
decorator==4.4.1
entrypoints==0.3
flake8==3.7.9
gast==0.3.2
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
jupyter-client==5.3.4
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
mccabe==0.6.1
numpy==1.18.0
opencv-python==4.1.2.30
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.2
protobuf==3.11.2
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
pynput==1.6.5
pyparsing==2.4.6
python-dateutil==2.8.1
python-xlib==0.26
pyzmq==18.1.1
six==1.13.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.8
Werkzeug==0.16.0
wrapt==1.11.2
